requirements.txt
pydantic-ai
pypdf2
markdown-it-py
python-dotenv
pytest
click
langchain
openai
pytest-asyncio
